se cambiaron controladores para recibir imagenes y datos json se elimino el archivo home user controller se creo un dto para el regidstro del mantenimiento se configuro una anotacion de json manage en las entidades  para que no se crearan ciclos infinitos en la respestas de json se modificaron sol servicios para registrar los mantenimientos tambien para tomar el mantenimiento con el id del vehiculo

+++ src/main/java/com/gestion/automange/service/IUsuarioService.java
@@ -3,9 +3,6 @@ package com.gestion.automange.service;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.security.core.userdetails.UserDetails;
-
-import com.gestion.automange.model.Productos;
 import com.gestion.automange.model.Usuario;
 
 public interface IUsuarioService {
@@ -18,7 +15,7 @@ public interface IUsuarioService {
 
 	public void delete(Integer id);
 
-	Optional<Usuario> findById(Integer id);
+	Optional<Usuario> findById(Long id);
 
 	Optional<Usuario> findByEmail(String email);
